Move some bits from org.sonatype.nexus.util to org.sonatype.nexus.proxy.utils, as they deal with legacy repository framework shits
